requirements.txt
fastapi
uvicorn[standard]
python-gitlab
pydantic-settings
